requirements.txt
boto3==1.26.137
pandas==1.5.3
psycopg2-binary==2.9.6
pyarrow==12.0.1